# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "data-loom.",
        "title": "data-loom API",
        "contact": {},
        "version": "v1.0"
    },
    "host": "localhost:8080",
    "basePath": "/data-loom/api/v1",
    "paths": {
...
        "/task": {
            "get": {
                "description": "查询任务",
                "tags": [
                    "任务"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建人",
                        "name": "creator",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改人",
                        "name": "editor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "描述",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否生效",
                        "name": "enable",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "策略ID",
                        "name": "policy_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "计划执行日期: 2024-03-01",
                        "name": "execute_date",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "执行窗口外是否需要暂停执行",
                        "name": "pause",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "执行窗口外是否重建表(仅在治理方式是删除时有效)。true:在执行窗口外仍然走重建流程; false:执行窗口外跳过重建流程",
                        "name": "rebuild_flag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "任务状态",
                        "name": "task_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务失败原因",
                        "name": "task_reason",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务失败详情",
                        "name": "task_detail",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "工作流",
                        "name": "workflow",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "源端ID",
                        "name": "src_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "源端名称",
                        "name": "src_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "资产BU",
                        "name": "src_bu",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "源端集群名称",
                        "name": "src_cluster_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "源端集群ID",
                        "name": "src_cluster_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "源库名",
                        "name": "src_database_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "源表名",
                        "name": "src_tables_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "源端归档列名",
                        "name": "src_columns",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "truncate",
                            "delete",
                            "backup-delete",
                            "archive"
                        ],
                        "type": "string",
                        "description": "数据治理方式",
                        "name": "govern",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "数据治理条件",
                        "name": "condition",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否清理源表",
                        "name": "clean_src",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "steady",
                            "balanced",
                            "swift"
                        ],
                        "type": "string",
                        "description": "清理速度",
                        "name": "cleaning_speed",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "目标端ID",
                        "name": "dest_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "目标端名称",
                        "name": "dest_name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "mysql",
                            "databend"
                        ],
                        "type": "string",
                        "description": "存储介质",
                        "name": "dest_storage",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "目标端连接ID",
                        "name": "dest_connection_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "目标端数据库",
                        "name": "dest_database_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "目标端表名字",
                        "name": "dest_table_name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "目标端是否压缩存储存储",
                        "name": "dest_compress",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关注人",
                        "name": "relevant",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "silence",
                            "success",
                            "failed",
                            "always"
                        ],
                        "type": "string",
                        "description": "通知策略",
                        "name": "notify_policy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.TaskService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新任务",
                "tags": [
                    "任务"
                ],
                "parameters": [
                    {
                        "description": "Task",
                        "name": "Task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.TaskService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.TaskService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除任务",
                "tags": [
                    "任务"
                ],
                "parameters": [
                    {
                        "description": "Task",
                        "name": "Task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.TaskService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.TaskService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/task/result": {
            "put": {
                "description": "上报任务执行结果",
                "tags": [
                    "任务"
                ],
                "parameters": [
                    {
                        "description": "TaskResult",
                        "name": "TaskResult",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.TaskResultService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.TaskService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.CleaningSpeedType": {
            "type": "string",
            "enum": [
                "steady",
                "balanced",
                "swift",
                "balanced"
            ],
            "x-enum-comments": {
                "CleaningSpeed": "速度优先",
                "CleaningSpeedBalanced": "速度适中",
                "CleaningSpeedSteady": "稳定优先"
            },
            "x-enum-varnames": [
                "CleaningSpeedSteady",
                "CleaningSpeedBalanced",
                "CleaningSpeed",
                "CleaningSpeedDefault"
            ]
        },
...
        "common.GovernType": {
            "type": "string",
            "enum": [
                "truncate",
                "delete",
                "backup-delete",
                "archive",
                "rebuild"
            ],
            "x-enum-comments": {
                "GovernTypeArchive": "归档数据",
                "GovernTypeBackupDelete": "备份后删除",
                "GovernTypeDelete": "删除数据",
                "GovernTypeRebuild": "重建表",
                "GovernTypeTruncate": "清空数据"
            },
            "x-enum-varnames": [
                "GovernTypeTruncate",
                "GovernTypeDelete",
                "GovernTypeBackupDelete",
                "GovernTypeArchive",
                "GovernTypeRebuild"
            ]
        },
...
        "common.NotifyPolicyType": {
            "type": "string",
            "enum": [
                "silence",
                "success",
                "failed",
                "always",
                "always"
            ],
            "x-enum-comments": {
                "NotifyPolicyTypeAlways": "成功或失败都通知",
                "NotifyPolicyTypeFailed": "失败时通知",
                "NotifyPolicyTypeSilence": "不通知",
                "NotifyPolicyTypeSuccess": "成功时通知"
            },
            "x-enum-varnames": [
                "NotifyPolicyTypeSilence",
                "NotifyPolicyTypeSuccess",
                "NotifyPolicyTypeFailed",
                "NotifyPolicyTypeAlways",
                "NotifyPolicyTypeDefault"
            ]
        },
...
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "返回码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据"
                },
                "error": {
                    "description": "错误信息",
                    "type": "string"
                },
                "message": {
                    "description": "返回信息",
                    "type": "string"
                }
            }
        },
        "common.StorageType": {
            "type": "string",
            "enum": [
                "mysql",
                "databend"
            ],
            "x-enum-varnames": [
                "StorageMysql",
                "StorageDataBend"
            ]
        },
...
        "services.TaskService": {
            "type": "object",
            "properties": {
                "cleaning_speed": {
                    "description": "清理速度 稳定优先:steady, 速度适中:balanced, 速度优先:swift",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.CleaningSpeedType"
                        }
                    ]
                },
                "condition": {
                    "description": "数据治理条件",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建人",
                    "type": "string"
                },
                "description": {
                    "description": "说明",
                    "type": "string"
                },
                "dest_compress": {
                    "description": "是否压缩存储",
                    "type": "boolean"
                },
                "dest_connection_id": {
                    "description": "归档库连接信息",
                    "type": "integer"
                },
                "dest_database_name": {
                    "description": "归档库名",
                    "type": "string"
                },
                "dest_id": {
                    "description": "目标端信息",
                    "type": "integer"
                },
                "dest_name": {
                    "description": "目标端名称",
                    "type": "string"
                },
                "dest_storage": {
                    "description": "归档介质",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.StorageType"
                        }
                    ]
                },
                "dest_table_name": {
                    "description": "归档表名",
                    "type": "string"
                },
                "editor": {
                    "description": "修改人",
                    "type": "string"
                },
                "enable": {
                    "description": "是否生效",
                    "type": "boolean"
                },
                "execute_date": {
                    "description": "预计执行日期 \"2024-01-02\"",
                    "type": "string"
                },
                "execute_window": {
                    "description": "执行窗口",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "govern": {
                    "description": "数据治理方式",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.GovernType"
                        }
                    ]
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "name": {
                    "description": "任务名称",
                    "type": "string"
                },
                "notify_policy": {
                    "description": "通知策略 不通知:silence, 成功时通知:success, 失败时通知:failed, 成功或失败都通知:always",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.NotifyPolicyType"
                        }
                    ]
                },
                "pause": {
                    "description": "执行窗口外是否需要暂停执行",
                    "type": "boolean"
                },
                "policy_id": {
                    "description": "策略ID",
                    "type": "integer"
                },
                "rebuild_flag": {
                    "description": "执行窗口外是否重建表(仅在治理方式是删除时有效)。true:在执行窗口外仍然走重建流程; false:执行窗口外跳过重建流程",
                    "type": "boolean"
                },
                "relevant": {
                    "description": "结果通知",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "retain_src_data": {
                    "description": "归档时否保留源表数据",
                    "type": "boolean"
                },
                "src_addr": {
                    "description": "源端地址",
                    "type": "string"
                },
                "src_bu": {
                    "description": "资产BU",
                    "type": "string"
                },
                "src_cluster_id": {
                    "description": "集群ID",
                    "type": "string"
                },
                "src_cluster_name": {
                    "description": "集群名称",
                    "type": "string"
                },
                "src_columns": {
                    "description": "源列名",
                    "type": "string"
                },
                "src_database_name": {
                    "description": "源库名",
                    "type": "string"
                },
                "src_id": {
                    "description": "源端信息",
                    "type": "integer"
                },
                "src_name": {
                    "description": "源端名称",
                    "type": "string"
                },
                "src_tables_name": {
                    "description": "源表名",
                    "type": "string"
                },
                "task_detail": {
                    "description": "任务失败详情",
                    "type": "string"
                },
                "task_duration": {
                    "description": "执行时长(秒)",
                    "type": "integer"
                },
                "task_end_time": {
                    "description": "执行结束时间",
                    "type": "string"
                },
                "task_reason": {
                    "description": "任务失败原因",
                    "type": "string"
                },
                "task_result_quantity": {
                    "description": "治理数据量",
                    "type": "integer"
                },
                "task_result_size": {
                    "description": "治理数据大小",
                    "type": "integer"
                },
                "task_start_time": {
                    "description": "开始执行时间",
                    "type": "string"
                },
                "task_status": {
                    "description": "任务状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.TaskStatusType"
                        }
                    ]
                },
                "updated_at": {
                    "description": "修改时间",
                    "type": "string"
                },
                "workflow": {
                    "description": "工作流",
                    "type": "string"
                },
                "workflow_url": {
                    "description": "工作流地址",
                    "type": "string"
                }
            }
        },
...
        "types.TaskResultService": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "任务ID",
                    "type": "integer"
                },
                "task_detail": {
                    "description": "任务失败详情",
                    "type": "string"
                },
                "task_end_time": {
                    "description": "执行结束时间 \"2006-01-02 15:04:05\"",
                    "type": "string"
                },
                "task_reason": {
                    "description": "任务失败原因",
                    "type": "string"
                },
                "task_result_quantity": {
                    "description": "治理数据量",
                    "type": "integer"
                },
                "task_result_size": {
                    "description": "治理数据容量",
                    "type": "integer"
                },
                "task_start_time": {
                    "description": "开始执行时间 \"2006-01-02 15:04:05\"",
                    "type": "string"
                },
                "task_status": {
                    "description": "任务状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.TaskStatusType"
                        }
                    ]
                }
            }
        }
    }
}